Wrote unit tests for pixel path algorithm.
